feat: stabiliseringspakke — /ready, domene-klienter, apiConfig, test-scenarios-cleanup (#43,#46,#48,#50,#65)
- GET /ready med DB-tilkoblingssjekk og migrasjonsvalidering (readiness.ts + test)
- localServiceClient.ts splittet i 7 domene-klienter; barrel beholder bakoverkompatibilitet
- apiConfig.ts sentraliserer NEXT_PUBLIC_API_BASE_URL; ChargenWizard og klienter bruker den
- @glantri/test-scenarios fjernet fra packages/database produksjonsavhengigheter
- scenarios.ts: access.ts og parsing.ts løftet ut som første trygge splitting
- workbookCombatMath.test.ts utvidet med 178 linjer nye golden tests
- CI/deploy bruker /ready i smoke-sjekk
