+++ .github/workflows/ci.yml
@@ -109,9 +109,9 @@ jobs:
       - name: Wait for API to be ready
         run: |
           for i in $(seq 1 12); do
-            if curl -fsS http://localhost:4000/health; then
+            if curl -fsS http://localhost:4000/ready; then
               echo ""
-              echo "Health check passed on attempt $i"
+              echo "Readiness check passed on attempt $i"
               exit 0
             fi
             echo "Attempt $i/12 failed, retrying in 5s..."
