+++ packages/database/package.json
@@ -31,7 +31,6 @@
     "@glantri/domain": "workspace:*",
     "@glantri/rules-engine": "workspace:*",
     "@glantri/shared": "workspace:*",
-    "@glantri/test-scenarios": "workspace:*",
     "@prisma/client": "^6.6.0"
   },
   "devDependencies": {
